[search] Fixed upper bound on number of open fds for features collector tool.
